Pretty

Description

module that contains class and instances of Pretty

Documentation

class Show a => Pretty a where

Instances of this class provide the functions pretty which creates a string of this object. By deafualt it ist used show.

Minimal complete definition

Nothing

Methods

pretty :: a -> String

takes the object and returns a representation of the string that looks nicely.