Strategy

Description

module to solve a Goal with given program by choosing search algorithm in SLD Tree

Documentation

dfs :: Strategy

depth first search in sld tree

bfs :: Strategy

breadth first search in sld tree

solve :: Strategy -> Prog -> Goal -> [Subst]

create the sld tree given a program and a goal to proof and search in the sld tree with the search strategy also given