Subst

Description

module for substitutions. apply and compose substitions

constructors

empty :: Subst

creates an empty substitution

single

Arguments

:: VarIndex
the variable index to be replaced

-> Term
the term the variable is replaced with

-> Subst

creates a substitution that has only one variable exchange

applications

apply

Arguments

:: Subst
the subtitution to apply

-> Term
the term the substitution is applied to

-> Term

applies a substitution to a given term and replaces all the variables in the term with the terms given in the subtitution

compose

Arguments

:: Subst
substitution s2

-> Subst
substitution s1

-> Subst

compose to substitutions with each other. the returned substitution has the following elements: let a -> b be a variable exchange of the first substitution (s1) then the goal substitution contains: a -> apply s2 b, when s2 is the second substitution. the goal substitution also contains all c -> d of s2 for all c that are not present in s1 on the left side.